Refactoring: Moved BaseJettyServerInitializer.java to JettyServerInitUtils.java and changed it from abstract class to a utils class containing static methods And, moved jetty stuff to its own package io.druid.server.initialization.jetty
