+++ server/src/main/java/io/druid/server/initialization/jetty/JettyServerInitUtils.java
@@ -1,21 +1,23 @@
 /*
- * Druid - a distributed column store.
- * Copyright 2012 - 2015 Metamarkets Group Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+* Licensed to Metamarkets Group Inc. (Metamarkets) under one
+* or more contributor license agreements. See the NOTICE file
+* distributed with this work for additional information
+* regarding copyright ownership. Metamarkets licenses this file
+* to you under the Apache License, Version 2.0 (the
+* "License"); you may not use this file except in compliance
+* with the License. You may obtain a copy of the License at
+*
+* http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing,
+* software distributed under the License is distributed on an
+* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+* KIND, either express or implied. See the License for the
+* specific language governing permissions and limitations
+* under the License.
+*/
 
-package io.druid.server.initialization;
+package io.druid.server.initialization.jetty;
 
 import java.util.Set;
 
@@ -32,19 +34,19 @@ import org.eclipse.jetty.servlets.GzipFilter;
 
 import javax.ws.rs.HttpMethod;
 
-public abstract class BaseJettyServerInitializer implements JettyServerInitializer
+public class JettyServerInitUtils
 {
 
   public static final String GZIP_METHODS = Joiner.on(",").join(HttpMethod.GET, HttpMethod.POST);
 
-  public FilterHolder defaultGzipFilterHolder()
+  public static FilterHolder defaultGzipFilterHolder()
   {
     final FilterHolder gzipFilterHolder = new FilterHolder(GzipFilter.class);
     setDefaultGzipFilterHolderParameters(gzipFilterHolder);
     return gzipFilterHolder;
   }
 
-  public FilterHolder defaultAsyncGzipFilterHolder()
+  public static FilterHolder defaultAsyncGzipFilterHolder()
   {
     final FilterHolder gzipFilterHolder = new FilterHolder(AsyncGzipFilter.class);
     setDefaultGzipFilterHolderParameters(gzipFilterHolder);
@@ -60,7 +62,7 @@ public abstract class BaseJettyServerInitializer implements JettyServerInitializ
     filterHolder.setInitParameter("checkGzExists", String.valueOf(false));
   }
 
-  public void addExtensionFilters(ServletContextHandler handler, Injector injector) {
+  public static void addExtensionFilters(ServletContextHandler handler, Injector injector) {
     Set<ServletFilterHolder> extensionFilters = injector.getInstance(Key.get(new TypeLiteral<Set<ServletFilterHolder>>(){}));
 
     for (ServletFilterHolder servletFilterHolder : extensionFilters) {
