+++ services/src/main/java/io/druid/cli/CliBridge.java
@@ -54,8 +54,9 @@ import io.druid.server.coordination.AbstractDataSegmentAnnouncer;
 import io.druid.server.coordination.BatchDataSegmentAnnouncer;
 import io.druid.server.coordination.DruidServerMetadata;
 import io.druid.server.initialization.BatchDataSegmentAnnouncerConfig;
-import io.druid.server.initialization.JettyServerInitializer;
 import io.druid.server.initialization.ZkPathsConfig;
+import io.druid.server.initialization.jetty.JettyServerInitializer;
+
 import org.apache.curator.framework.CuratorFramework;
 import org.apache.curator.framework.CuratorFrameworkFactory;
 import org.apache.curator.retry.BoundedExponentialBackoffRetry;
