+++ services/src/main/java/io/druid/cli/QueryJettyServerInitializer.java
@@ -19,7 +19,10 @@ package io.druid.cli;
 
 import com.google.inject.Injector;
 import com.google.inject.servlet.GuiceFilter;
-import io.druid.server.initialization.BaseJettyServerInitializer;
+
+import io.druid.server.initialization.jetty.JettyServerInitUtils;
+import io.druid.server.initialization.jetty.JettyServerInitializer;
+
 import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.handler.HandlerList;
@@ -29,15 +32,15 @@ import org.eclipse.jetty.servlet.ServletHolder;
 
 /**
 */
-public class QueryJettyServerInitializer extends BaseJettyServerInitializer
+public class QueryJettyServerInitializer implements JettyServerInitializer
 {
   @Override
   public void initialize(Server server, Injector injector)
   {
     final ServletContextHandler root = new ServletContextHandler(ServletContextHandler.SESSIONS);
     root.addServlet(new ServletHolder(new DefaultServlet()), "/*");
-    addExtensionFilters(root, injector);
-    root.addFilter(defaultGzipFilterHolder(), "/*", null);
+    JettyServerInitUtils.addExtensionFilters(root, injector);
+    root.addFilter(JettyServerInitUtils.defaultGzipFilterHolder(), "/*", null);
 
     root.addFilter(GuiceFilter.class, "/*", null);
